requirements.txt
selenium==4.18.1
undetected-chromedriver==3.5.5
beautifulsoup4==4.12.3
requests==2.31.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
openpyxl==3.1.5
setuptools>=68.0.0
